backend/requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==21.2.0
bcrypt==4.1.3
blinker==1.8.2
blueprint==3.4.2
bytesbufio==1.0.3
cachelib==0.9.0
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
colorama==0.4.6
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
DateTime==5.5
dbus-python==1.2.18
diffusers==0.30.3
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.21.2
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
flask-blueprint==1.3.0
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-ngrok==0.0.25
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.0
fsspec==2024.6.1
gradio==5.9.1
gradio_client==1.5.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.27.0
idna==3.7
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==3.2.0
keyring==23.5.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kivymd==1.2.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.5
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
msgspec==0.18.6
mysql-connector-python==9.0.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
openai==1.37.1
opencv-python==4.10.0.84
orjson==3.10.13
packaging==24.1
pandas==2.2.3
peace==0.0.8
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.8.0
pymongo==4.9.1
PyMySQL==1.1.1
pyparsing==2.4.7
pypng==0.20220715.0
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==5.4.1
qrcode==7.4.2
redis==4.6.0
regex==2024.7.24
requests==2.32.3
rich==13.9.4
ruff==0.8.5
safehttpx==0.1.6
safetensors==0.4.3
scipy==1.14.1
SecretStorage==3.3.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.1
starlette==0.41.3
svgwrite==1.4.3
sympy==1.13.1
systemd-python==234
texttable==1.6.4
tk==0.1.0
tokenizers==0.20.0
tomlkit==0.13.2
torch==2.4.0
tqdm==4.66.4
transformers==4.45.1
triton==3.0.0
ttystatus==0.38
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
uvicorn==0.34.0
virtualenv==20.26.3
wadllib==1.3.6
waitress==3.0.0
watchdog==6.0.0
websocket-client==1.2.3
websockets==14.1
Werkzeug==3.0.3
zipp==1.0.0
zope.interface==6.4.post2
